requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
pandas
sqlmodel
psycopg2-binary
python-dotenv
python-multipart